master : Added Character namespace and scripts

+++ Assets/Scripts/CharacterSelection/CharacterSelectionSettings.cs
@@ -9,8 +9,7 @@ namespace Game.CharacterSelection
     {
         public int CrewSize;
         public int CharactersPerSelection;
-        public float SecondsPerSelection;
+        public int SecondsPerSelection;
         public GameObject CharacterPrefab;
     }
 }
-
